Remove extension & circular ref issue
