+++ src/Easify.Extensions.Notifications.Extensions/Easify.Extensions.Notifications.Extensions.csproj
@@ -18,7 +18,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\Easify.Extensions.Notifications.MessagingService\Easify.Extensions.Notifications.MessagingService.csproj" />
       <ProjectReference Include="..\Easify.Extensions.Notifications\Easify.Extensions.Notifications.csproj" />
-      <ProjectReference Include="..\Easify.Extensions.Notifications.Messaging\Easify.Extensions.Notifications.Messaging.csproj" />
     </ItemGroup>
 </Project>
