+++ src/Easify.Extensions.Notifications.Messaging/Configuration/NotificationAudience.cs
@@ -1,4 +1,4 @@
-// This software is part of the Easify framework
+﻿// This software is part of the Easify framework
 // Copyright (C) 2019 Intermediate Capital Group
 // 
 // This program is free software: you can redistribute it and/or modify
@@ -14,14 +14,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-
-namespace Easify.Extensions.Notifications.Exceptions
+namespace Easify.Extensions.Notifications.Messaging.Configuration
 {
-    public class SmtpOptionsException : Exception
+    public sealed class NotificationAudience
     {
-        public SmtpOptionsException(string propertyName) : base($"{propertyName} is missing from the configuration")
-        {
-        }
+        public string Email { get; set; }
     }
 }
