+++ src/Easify.Extensions.Notifications.Messaging/Configuration/NotificationTemplate.cs
@@ -14,10 +14,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Easify.Extensions.Notifications.Configuration
+namespace Easify.Extensions.Notifications.Messaging.Configuration
 {
-    public static class NotificationProfileNames
+    public class NotificationTemplate
     {
-        public const string DefaultProfile = "DefaultProfile";
+        public string Name { get; set; }
+        public string Path { get; set; }
     }
 }
