+++ src/Easify.Extensions.Notifications.Messaging/Easify.Extensions.Notifications.Messaging.csproj
@@ -2,24 +2,10 @@
 
     <PropertyGroup>
         <TargetFramework>netstandard2.0</TargetFramework>
-        <Description>Notification Messaging</Description>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-        <Authors>Mohammad Moattar</Authors>
-        <Company>ICG</Company>
-        <PackageProjectUrl>http://github.com/icgam/Easify</PackageProjectUrl>
-        <RepositoryUrl>https://github.com/icgam/Easify</RepositoryUrl>
-        <GenerateAssemblyVersionAttribute>false</GenerateAssemblyVersionAttribute>
-        <Version>1.0.0</Version>
-        <LangVersion>8</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="MailKit" Version="2.10.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Easify.Extensions.Notifications\Easify.Extensions.Notifications.csproj" />
-      <ProjectReference Include="..\Easify.Extensions.Notifications\Easify.Extensions.Notifications.csproj" />
+      <PackageReference Include="FluentValidation" Version="9.2.0" />
     </ItemGroup>
 
 </Project>
