+++ src/Easify.Extensions.Notifications.Messaging/EmailAddress.cs
