+++ src/Easify.Extensions.Notifications.Messaging/Message.cs
