+++ src/Easify.Extensions.Notifications.sln
@@ -1,8 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications.Extensions", "Easify.Extensions.Notifications.Extensions\Easify.Extensions.Notifications.Extensions.csproj", "{8B996EE1-1A29-4991-8512-65DE30D90651}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications.Messaging", "Easify.Extensions.Notifications.Messaging\Easify.Extensions.Notifications.Messaging.csproj", "{A75412D9-6464-46E2-84ED-5D90D1DE3A89}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications.MessagingService", "Easify.Extensions.Notifications.MessagingService\Easify.Extensions.Notifications.MessagingService.csproj", "{A75412D9-6464-46E2-84ED-5D90D1DE3A89}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications.Templating", "Easify.Extensions.Notifications.Templating\Easify.Extensions.Notifications.Templating.csproj", "{AA2C8AD1-8CDB-4AFF-8B40-AD820A24C418}"
 EndProject
@@ -14,16 +12,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notificat
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications", "Easify.Extensions.Notifications\Easify.Extensions.Notifications.csproj", "{60381FF9-B2D5-4B94-95BC-E07252C0620B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Easify.Extensions.Notifications.Messaging", "Easify.Extensions.Notifications.Messaging\Easify.Extensions.Notifications.Messaging.csproj", "{CD8000BB-78D2-4BAF-9434-2C3A43E76DE7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8B996EE1-1A29-4991-8512-65DE30D90651}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B996EE1-1A29-4991-8512-65DE30D90651}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B996EE1-1A29-4991-8512-65DE30D90651}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B996EE1-1A29-4991-8512-65DE30D90651}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A75412D9-6464-46E2-84ED-5D90D1DE3A89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A75412D9-6464-46E2-84ED-5D90D1DE3A89}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A75412D9-6464-46E2-84ED-5D90D1DE3A89}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -48,5 +44,9 @@ Global
 		{60381FF9-B2D5-4B94-95BC-E07252C0620B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{60381FF9-B2D5-4B94-95BC-E07252C0620B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{60381FF9-B2D5-4B94-95BC-E07252C0620B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CD8000BB-78D2-4BAF-9434-2C3A43E76DE7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD8000BB-78D2-4BAF-9434-2C3A43E76DE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD8000BB-78D2-4BAF-9434-2C3A43E76DE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD8000BB-78D2-4BAF-9434-2C3A43E76DE7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
